Split TextArea stories
