--- src/js/components/TextArea/textarea.stories.js
@@ -1,71 +1,0 @@
-import React from 'react';
-import { storiesOf } from '@storybook/react';
-
-import { Grommet, Box, TextArea } from 'grommet';
-import { grommet } from 'grommet/themes';
-import { deepMerge } from 'grommet/utils';
-
-const SimpleTextArea = props => {
-  const [value, setValue] = React.useState('');
-
-  const onChange = event => setValue(event.target.value);
-
-  return (
-    <Grommet theme={grommet}>
-      <Box align="center" pad="large">
-        <TextArea value={value} onChange={onChange} {...props} />
-      </Box>
-    </Grommet>
-  );
-};
-
-const customTheme = deepMerge(grommet, {
-  textArea: {
-    extend: () => `
-      font-size: 40px;
-      color: red;
-    `,
-  },
-});
-
-const ThemedTextArea = () => {
-  const [value, setValue] = React.useState('');
-
-  const onChange = event => setValue(event.target.value);
-
-  return (
-    <Grommet theme={customTheme}>
-      <Box
-        width="large"
-        height="medium"
-        border={{ color: 'brand', size: 'medium' }}
-      >
-        <TextArea value={value} onChange={onChange} fill />
-      </Box>
-    </Grommet>
-  );
-};
-
-const FillTextArea = () => {
-  const [value, setValue] = React.useState('');
-
-  const onChange = event => setValue(event.target.value);
-
-  return (
-    <Grommet theme={grommet}>
-      <Box
-        width="large"
-        height="medium"
-        border={{ color: 'brand', size: 'medium' }}
-      >
-        <TextArea value={value} onChange={onChange} fill />
-      </Box>
-    </Grommet>
-  );
-};
-
-storiesOf('TextArea', module)
-  .add('Simple', () => <SimpleTextArea resize />)
-  .add('Fill', () => <FillTextArea />)
-  .add('Themed', () => <ThemedTextArea />)
-  .add('Non resizable', () => <SimpleTextArea resize={false} />);
